fix for loops
